debug

+++ OI/USACO/2.4.6.cpp
@@ -23,7 +23,7 @@ int main()
             m/=i;
         }
     }
-    double ans=n/m;
+    double ans=(double)n/(double)m;
     cout<<ans;
     return 0;
 }
